fix(tmux): Prevent printing escape sequences in VS Code on startup
The offending config item was "set -sg escape-time 0". According to the
manual, this determines how long tmux "waits after an escape is input to
determine if it is part of a function or meta key sequences".
Presumably, this caused tmux to misinterpret an escape sequence.
Increasing it to 100ms appears to have done the trick.